requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
lxml==4.3.4
numpy==1.17.2
pandas==0.25.1
Pillow==6.0.0
psycopg2==2.8.3
PyQt5==5.12.2
PyQt5-sip==4.19.17
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.2
six==1.12.0
toml==0.10.0
xlrd==1.2.0
XlsxWriter==1.1.8
